Updated all tests that do file I/O so that they work in parallel.  - In most cases, the I/O checking was restricted to processor 0  - However in the case of the Journal, I/O Streams is deliberately done on all processors, to test    it works.

+++ Base/IO/tests/PathUtilsSuite.c
@@ -34,6 +34,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <mpi.h>
 
 #include "pcu/pcu.h"
 #include "StGermain/Base/Foundation/Foundation.h"
@@ -41,10 +42,12 @@
 #include "PathUtilsSuite.h"
 
 typedef struct {
+   unsigned int   rank;
 } PathUtilsSuiteData;
 
 
 void PathUtilsSuite_Setup( PathUtilsSuiteData* data ) {
+   MPI_Comm_rank( MPI_COMM_WORLD, &data->rank );
 }
 
 void PathUtilsSuite_Teardown( PathUtilsSuiteData* data ) {
@@ -74,42 +77,50 @@ void PathUtilsSuite_TestFindFile( PathUtilsSuiteData* data ) {
    char        fullPath[1024];
    const char* subDir = "./testSubDir";
    const char* subDirFilename = "./testSubDir/subDirTest.xml";
-   FILE*       subDirFile = NULL;
    const char* currDirFilename = "./currDirTest.xml";
-   FILE*       currDirFile = NULL;
 
 
    Stg_asprintf( &searchPaths, ".:%s:/does/not/exist", subDir );
    /* Create necessary test files/dirs */
-   currDirFile = fopen( currDirFilename, "w" );
-   fputs( "test.\n", currDirFile );
-   fclose( currDirFile );
-   mkdir( subDir, 0755 );
-   subDirFile = fopen( subDirFilename, "w" );
-   fputs( "test.\n", subDirFile );
-   fclose( subDirFile );
+   if (data->rank==0) {
+      FILE*       subDirFile = NULL;
+      FILE*       currDirFile = NULL;
+
+      currDirFile = fopen( currDirFilename, "w" );
+      fputs( "test.\n", currDirFile );
+      fclose( currDirFile );
+      mkdir( subDir, 0755 );
+      subDirFile = fopen( subDirFilename, "w" );
+      fputs( "test.\n", subDirFile );
+      fclose( subDirFile );
+   }
+   MPI_Barrier(MPI_COMM_WORLD);
 
    /* try and open some files using the search path */
+   /* Only do this using proc 0 - for why, see warning in Doxygen comment for the function. */
+   if (data->rank==0) {
+      /* This first test is to make sure it can handle files preceded with ./ */
+      FindFile( fullPath, currDirFilename, searchPaths );
+      pcu_check_true( 0 == strcmp( fullPath, currDirFilename ));
 
-   /* This first test is to make sure it can handle files preceded with ./ */
-   FindFile( fullPath, currDirFilename, searchPaths );
-   pcu_check_true( 0 == strcmp( fullPath, currDirFilename ));
+      FindFile( fullPath, "currDirTest.xml", searchPaths );
+      pcu_check_true( 0 == strcmp( fullPath, currDirFilename ));
+      
+      FindFile( fullPath, "subDirTest.xml", searchPaths );
+      pcu_check_true( 0 == strcmp( fullPath, subDirFilename ));
+      
+      FindFile( fullPath, "nofile.man", searchPaths );
+      pcu_check_true( 0 == strcmp( fullPath, "" ));
+      
+      FindFile( fullPath, "/Users/luke/Projects/StGermain/env_vars", searchPaths );
+      pcu_check_true( 0 == strcmp( fullPath, "" ));
+   }
 
-   FindFile( fullPath, "currDirTest.xml", searchPaths );
-   pcu_check_true( 0 == strcmp( fullPath, currDirFilename ));
-   
-   FindFile( fullPath, "subDirTest.xml", searchPaths );
-   pcu_check_true( 0 == strcmp( fullPath, subDirFilename ));
-   
-   FindFile( fullPath, "nofile.man", searchPaths );
-   pcu_check_true( 0 == strcmp( fullPath, "" ));
-   
-   FindFile( fullPath, "/Users/luke/Projects/StGermain/env_vars", searchPaths );
-   pcu_check_true( 0 == strcmp( fullPath, "" ));
-
-   remove( currDirFilename );
-   remove( subDirFilename );
-   rmdir( subDir );
+   if (data->rank==0) {
+      remove( currDirFilename );
+      remove( subDirFilename );
+      rmdir( subDir );
+   }
 }
